Refactor: Increase maintainability in codebase
- Moved all functions related to background operations to background.js for better modularity.
- Moved all functions related to key events to keys.js to separate concerns and improve readability.
- Moved all functions related to mouse events to mouse.js to enhance code organization and maintainability.
