+++ index.html
@@ -210,14 +210,14 @@
                     <p>
                         If you enjoyed exploring the projects, please consider leaving a
                         <a href="https://github.com/Vpekdas/Vpekdas.github.io" target="_blank">
-                            <img src="./Star_Sharp_Solid.svg" alt="Star Icon" class="star-icon" />
+                            <img src="/ui/Star_Sharp_Solid.svg" alt="Star Icon" class="star-icon" />
                         </a>
                         on GitHub.
                     </p>
                     <p>
                         Ready for another round? Click the button below to restart your adventure:
                         <button id="reset-button">
-                            <img src="./Star_Classic_Solid.svg" alt="Reset Icon" class="reset-icon" />
+                            <img src="/ui/Star_Classic_Solid.svg" alt="Reset Icon" class="reset-icon" />
                         </button>
                     </p>
                 </div>
@@ -227,9 +227,9 @@
             <div id="phone-frame">
                 <div id="phone-screen">
                     <div id="status-bar">
-                        <img src="bar-signal.png" class="bar-signal" />
-                        <img src="email-icon.png" class="mail-icon" />
-                        <img src="battery.png" class="battery-icon" />
+                        <img src="/phone/bar-signal.png" class="bar-signal" />
+                        <img src="/phone/email-icon.png" class="mail-icon" />
+                        <img src="/phone/battery.png" class="battery-icon" />
                     </div>
                     <div id="dmail-date">
                         <span>受信日時: 2023.12.12</span>
@@ -237,7 +237,7 @@
                     </div>
                     <div id="dmail-from">
                         <span>差出人: </span>
-                        <span id="sender">Okabe Rintarou</span>
+                        <span id="sender">Okabe Rintaro</span>
                     </div>
                     <div id="dmail-subject">
                         <span>件名: Time travel</span>
