+++ public/phone/bar-signal.png
