+++ public/steins-gate/spritesheet.png
